# api/openapi.yaml
openapi: 3.0.3
info:
  title: Finny API
  description: Personal finance management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
  /transactions:
    post:
      operationId: createTransaction
      summary: Create a transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listTransactions
      summary: List transactions
      tags: [Transactions]
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: start_date
          in: query
          schema:
            type: string
            format: date
            example: '2024-01-01'
        - name: end_date
          in: query
          schema:
            type: string
            format: date
            example: '2024-12-31'
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/TransactionID'
    get:
      operationId: getTransaction
      summary: Get a transaction
      tags: [Transactions]
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteTransaction
      summary: Delete a transaction
      tags: [Transactions]
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateTransaction
      summary: Update a transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /transactions/{id}/status:
    parameters:
      - $ref: '#/components/parameters/TransactionID'
    patch:
      operationId: updateTransactionStatus
      summary: Update transaction status
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '204':
          description: Status updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    TransactionID:
      name: id
      in: path
      required: true
      description: Transaction UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    TransactionType:
      type: string
      enum: [income, expense]
    TransactionStatus:
      type: string
      enum: [draft, pending_invoice, complete, no_invoice]
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          description: Amount in cents
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          $ref: '#/components/schemas/TransactionStatus'
        description:
          type: string
        raw_description:
          type: string
        date:
          type: string
          format: date-time
        document_id:
          type: string
          format: uuid
          nullable: true
        document:
          allOf:
            - $ref: '#/components/schemas/Document'
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        mime_type:
          type: string
...
    CreateTransactionRequest:
      type: object
      required: [amount, type, description, date]
      properties:
        amount:
          type: integer
          format: int64
          description: Amount in cents, must not be zero
          example: 4999
        type:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
          example: Grocery shopping
        date:
          type: string
          format: date-time
          example: '2024-06-15T00:00:00Z'
    UpdateTransactionRequest:
      type: object
      properties:
        description:
          type: string
        amount:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/TransactionType'
        date:
          type: string
          format: date-time
        no_invoice:
          type: boolean
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/TransactionStatus'
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found